refactor(cli): split runtime infrastructure
