+++ packages/cli/src/types/wiggum-agent.d.ts
@@ -1,0 +1,7 @@
+declare module '@wiggum/agent' {
+  import type { Config } from '@opencode-ai/sdk';
+
+  export function buildMergedConfig(options?: {
+    fetchEnv?: () => unknown | Promise<unknown>;
+  }): Promise<Config>;
+}
