+++ packages/cli/tsconfig.json
@@ -3,6 +3,7 @@
     "target": "ES2020",
     "module": "Node16",
     "lib": ["ES2020"],
+    "baseUrl": ".",
     "outDir": "./dist",
     "rootDir": "./src",
     "strict": true,
